# Use AtomicReferenceArray to keep track of worker threads 