--- a/core/jvm/src/main/scala/cats/effect/unsafe/WorkerThread.scala
+++ b/core/jvm/src/main/scala/cats/effect/unsafe/WorkerThread.scala
@@ -553,7 +553,7 @@ private[effect] final class WorkerThread[P <: AnyRef](
             val idx = index
             val threadCount = pool.getWorkerThreadCount()
             val otherIdx = (idx + random.nextInt(threadCount - 1)) % threadCount
-            val thread = pool.getWorkerThreads(otherIdx)
+            val thread = pool.getWorkerThread(otherIdx)
             val state = thread.getState()
             val parked = thread.parked