For the most part the array of worker threads is read-only, however it may be updated from any thread when a worker thread transitions to blocking or back.

Previously, a separate atomic variable was used to introduce memory barriers when accessing the array. This is discarded in favour of `AtomicReferenceArray` that specifically manages access to elements of the array.

The `WorkStealingThreadPool.getWorkerThreads` method is replaced by a more limited `getWorkerThread(index: Int)` method. This is used by `WorkerThread` when checking for blocked threads, and the new method can be used to specifically access a single worker thread.

~*Note:* This PR includes the changes also proposed in #4195, but the intent is that they may be discussed separately. I can keep the two PRs in sync with regards to the common changes. In case the former PR is rejected I will drop the rejected changes here and otherwise update this one accordingly.~