@djspiewak 2024-12-21
This is very surprising to me. Arm chair reasoning really suggests that this should be a significant performance hit, but the benchmarks say it's a wash. Will think a bit more.

## Before

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score    Error    Units
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   10   724.626 ±  1.581    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   10    69.191 ±  0.095    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   10    11.697 ±  0.121  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   10    28.175 ±  6.114  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   10  2399.482 ± 51.270  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   10  1886.390 ±  5.104  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   10    22.989 ±  3.496  ops/min
```

## After

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score   Error    Units
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   10   722.919 ± 1.127    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   10    69.215 ± 0.234    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   10    11.858 ± 0.119  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   10    28.503 ± 7.477  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   10  2518.366 ± 4.768  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   10  1884.161 ± 5.135  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   10    25.316 ± 3.531  ops/min
```