@vasilmkd 2024-12-21
Given that the benchmarks are not showing any obvious regressions, the code is much simpler and clearer. Why not. 👍🏻